+++ src/scorer/feature/mentions_followed_scorer.ts
@@ -16,8 +16,8 @@ export default class MentionsFollowedScorer extends FeatureScorer {
 
     // Build simple dictionary of followed accounts (key is webfingerURI, value is 1)
     async prepareScoreData(): Promise<StringNumberDict> {
-        const followedAccounts = await MastoApi.instance.getFollowedAccounts();
-        return countValues<Account>(followedAccounts, (account) => account.webfingerURI);
+        // TODO: this is duplicative of the followedAccounts prop in UserData (wastes some memory, but not much)
+        return Account.buildWebfingerUriLookup(await MastoApi.instance.getFollowedAccounts())
     };
 
     // Toot.repair() already made StatusMention.acct fields equivalent to Account.webfingerURI