+++ src/scorer/feature/most_favorited_accounts_scorer.ts
@@ -1,9 +1,8 @@
 /*
  * Score how many times the current user has favourited the tooter in the past.
  */
+import Account from '../../api/objects/account';
 import AccountScorer from '../acccount_scorer';
-import Toot from '../../api/objects/toot';
-import { countValues } from '../../helpers/collection_helpers';
 import MastoApi from '../../api/api';
 import { StringNumberDict, WeightName } from '../../types';
 
@@ -15,6 +14,6 @@ export default class MostFavoritedAccountsScorer extends AccountScorer {
 
     async prepareScoreData(): Promise<StringNumberDict> {
         const recentFavourites = await MastoApi.instance.getRecentFavourites();
-        return countValues<Toot>(recentFavourites, (toot) => toot.account?.webfingerURI);
+        return Account.buildWebfingerUriLookup(recentFavourites.map(toot => toot.account));
     };
 };