+++ src/scorer/feature/most_retooted_users_scorer.ts
@@ -2,9 +2,8 @@
  * Score a toot based on how many times the user has retooted the author (or
  * the original author if it's a retoot).
  */
+import Account from '../../api/objects/account';
 import AccountScorer from "../acccount_scorer";
-import Toot from '../../api/objects/toot';
-import { countValues } from "../../helpers/collection_helpers";
 import MastoApi from "../../api/api";
 import { StringNumberDict, WeightName } from "../../types";
 
@@ -16,7 +15,7 @@ export default class MostRetootedUsersScorer extends AccountScorer {
 
     async prepareScoreData(): Promise<StringNumberDict> {
         const recentToots = await MastoApi.instance.getRecentUserToots();
-        const recentRetoots = recentToots.filter(toot => toot?.reblog);
-        return countValues<Toot>(recentRetoots, (toot) => toot.reblog?.account?.webfingerURI);
+        const retootedAccounts = recentToots.filter(toot => toot?.reblog).map(toot => toot.reblog!.account);
+        return Account.buildWebfingerUriLookup(retootedAccounts);
     };
 };