+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # NEXT RELEASE
 * Simplify the alternate score dict
 * Disallow setting the catchup checkpoint before the overall app timeline cutoff
+* UserData object now just stores a lookup dict for followedAccounts instead of all the actual followed Account objects
+* Refactor generic Account.buildWebfingerUriLookup() method
 
 ### v0.24.6
 * `MastoApi.instance.setBackgroundConcurrency()`, `Config.maxConcurrentRequestsBackground` option. Handle missing `Storage.getIdentity()` if possibl